+++ src/apis/referral_api.rs
@@ -65,7 +65,7 @@ pub struct ReferralControllerListReferralsParams {
     /// Direction to paginate through objects
     pub order: Option<String>,
     /// Limit the number of objects to return
-    pub limit: Option<i32>,
+    pub limit: Option<i64>,
     /// Pointer to the current object in pagination dataset
     pub cursor: Option<String>,
     /// Bytes32 encoded subaccount name (0x prefix, zero padded, set when using linked signer)
